change setup of QMAKE_MOC
don't load(moc) explicitly - turns out that this breaks
debug_and_release builds (the feature does not like being loaded before
default_post has been loaded).
so instead just do a minimal call to find the moc executable.

as it's in fact only two files which need it, so de-centralize this call
for the sake of saving some cpu cycles spent in qmake.

Change-Id: Ia0038f893c3931571e1494894bedcae85ebb4c8e
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>