[Qt] MediaPlayerQTKit broken after switching QPainter to raster engine. https://bugs.webkit.org/show_bug.cgi?id=104550
With Qt5 QPainter changed to using the raster engine.
Therefore we no longer have an NSGraphicsContext that we can
use for drawing video frames.
With this patch we are rendering the video frame into an image buffer
and then paint it manually using QPainter.

Reviewed by Simon Hausmann.

* platform/graphics/mac/MediaPlayerPrivateQTKit.mm:
(WebCore):
(WebCore::swapBgrToRgb):
     Swap the color format from BGR to RGB,
     since BGR is not supported by QImage.
(WebCore::MediaPlayerPrivateQTKit::paint):

Change-Id: I28462879baaa4c19b6fc642b38aadef5d9d3aa93
git-svn-id: http://svn.webkit.org/repository/webkit/trunk@137159 268f45cc-cd09-0410-ab3c-d52691b4dbf
Task-number: QTBUG-28361
Reviewed-by: Andras Becsi <andras.becsi@digia.com>