Revert r136231 to fix overlay size due to incorrect visible rect
This issue is visible when loading the full-sized image in the
flickrview example by clicking on a thumbnail.
Because of css device adaptation the code already changed in
trunk somewhat but the issue is also visible there, both on
retina and non-retina displays.
The problem here is that the visible rect that is set in
WebPage::sendViewportAttributesChanged() is incorrect because
it includes a rounding error if dpr != 1.0 since the stored
viewport size has integer precision, additionally there is also
a dependency problem since the constructed visible rect does
not include the UI-side scale that might be applied later in
page viewport controller due to the changed attributes.
This is especially visible on Flickr since the full size images
are shown in an overlay and an invalid visible rect results in
wrong overlay size.
In trunk this codepath is also used by css device adaptation
thus we might need a more comprehensive fix there, but since
we currently do not support css device adaptation we can revert
http://trac.webkit.org/changeset/136231 in this branch.

The issue is tracked upstream on:
https://bugs.webkit.org/show_bug.cgi?id=104906

Change-Id: Ia451ee610795700b0d3d6455413b97b607e6fb32
Reviewed-by: Andras Becsi <andras.becsi@digia.com>
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>