[Qt] Don't crash on exit when encountering uninitialized WebView.
https://bugs.webkit.org/show_bug.cgi?id=107812

When a WebView is never shown it will not initialize its root layer and
can thus crash on exit when trying to clean up said layers children.

This was fixed properly upstream in commit 18ac4c73a2 but according to
Jocelyn it cannot be cherry-picked easily due to dependencies on other
changes.

Change-Id: I58f5df8d3cd2756ac4f0b0c6a60b19c5c5ef27e8
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>