[Qt] WebKit fails to compile if EGL headers are not in default INCLUDEPATH
https://bugs.webkit.org/show_bug.cgi?id=111859

Patch by Floris Bos <bos@je-eigen-domein.nl> on 2013-03-12
Reviewed by Jocelyn Turcotte.

The 3D graphics code wants to include EGL header files.
But on some platforms such as the Raspberry Pi those are not in /usr/include
but in another folder.
Fix adds "egl" to CONFIG when OpenGL ES2 is used, so the right include
paths are added.

* WebCore.pri:

git-svn-id: http://svn.webkit.org/repository/webkit/trunk@145506 268f45cc-cd09-0410-ab3c-d52691b4dbfc

Change-Id: Ia7011b196ce985bd2293cc7556e3eee069c2eb50
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>