Workaround for MSVC 2012 64 bit optimizer bug.
Having optimization turned on for the VS 2012 build causes a crash
in RenderTableSection::paintObject. The workaround for this is to
turn optimization off for this method until a fix in the compiler
has been provided.

This is tracked in bug https://bugs.webkit.org/show_bug.cgi?id=112900
for the upstream version of WebKit.

Please be aware that this is a work around for QtWebKit and did not go
through the WebKit review process and is therefore not part of WebKit
trunk as of now.

Task-number: QTBUG-30117
Change-Id: I76139eaa4c0f07de2d90d197c0b07ef4a576e585
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>