[Qt][WK2] WebView's interactive property is not fully respected
https://bugs.webkit.org/show_bug.cgi?id=113066
https://trac.webkit.org/r147909

Reviewed by Jocelyn Turcotte.
WK2 sign-off by Benjamin Poulain.

The QML WebView inherits the "interactive" property from Flickable
which is true by default, and disables the interaction with the
Flickable if set to false.
Resulting from the design of the WebView panning and flicking is
disabled by Flickable but to be consistent we also need to disable
double-tap gestures and pinch gestures since they would trigger
scale and position changes.

Change-Id: I7879d7fa4bd2ccaf711dc44012905d49c9d7e8fd
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>