Coordinated Graphics: a long page is scaled vertically while loading.
https://bugs.webkit.org/show_bug.cgi?id=109645

Backported from http://trac.webkit.org/changeset/142837

Reviewed by Noam Rosenthal.

When loading http://www.w3.org/TR/xpath-datamodel/, Coordinated Graphics draws
vertically scaled contents. It is because there is the difference between the
size of a layer and the size of CoordinatedBackingStore.

Currently, LayerTreeRenderer notifies the size to CoordinatedBackingStore
at the moment of creating, updating and removing a tile. However, it is not
necessary to send tile-related messages when the size of layer is changed.
So this patch resets the size of CoordinatedBackingStore when receiving the
message that is created when the size is changed: SyncLayerState.

There is no current way to reliably test flicker issues.

Change-Id: Ied0dfa7e333b0c7d89c33684b05ca9e1dfebd155
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>