Fix testing if program exists in PATH
https://bugs.webkit.org/show_bug.cgi?id=116216

Patch by Jonathan Liu <net147@gmail.com> on 2013-05-16
Reviewed by Jocelyn Turcotte.

functions.prf incorrectly uses win_cmd_shell to test whether to
add .exe to the executable name when checking if it exists in
the PATH environment variable. This works correctly when
building using Windows Command Prompt but not under MSYS shell.

To resolve this, check the win32 platform scope instead of
win_cmd_shell so that it correctly adds .exe to the executable
name on Windows regardless of the shell used when testing if
it exists in PATH.

* qmake/mkspecs/features/functions.prf:

Change-Id: Ide88521e4922da89ce49dc4bccbce790e753804e
git-svn-id: http://svn.webkit.org/repository/webkit/trunk@150223 268f45cc-cd09-0410-ab3c-d52691b4dbfc
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>