Fix debug builds for Windows32 x86 bots.
Reviewed by Allan Sandfeld Jensen.

There was an out of memory problem when linking the debug builds
that had been solved on Linux and MinGW32 already by stripping
debug info from production builds.

This patch applies the same for MSVC builds on x86 hosts and enable the
debug symbols again for builds running on x86_64 hosts.

Change-Id: Id9ead480f243d8e48494f350a7a96b2bd2570c95
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@digia.com>