Do not accept all touch events
QtWebKit currently accepts all touch events to prevent the lack of accept
on a TouchBegin from preventing receiving TouchUpdate and TouchEnd.

This behavior prevents QGuiApplication from converting touch events to
mouse events, and makes the WebView widget impossible to interact with
by touch.

This patch modifies the behavior slighly so that QtWebKit only accepts
touch events if they are needed for the active document.

Task-number: QTBUG-31938
Change-Id: Iba80ebfaf31c93f82356e3f0622039120921172a
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>