Fix performance regression in HTML5 canvas stroke()
To solve corner case of stroking intersecting path with a transparent
color we switched the strokePath implementation to using QPainterPathStroker
in Qt 5. Unfortunately this appears to be 100-1000x slower than using
QPainter::strokePath.

This patch partially reverts the change and now only use QPainterPathStroker
when stroking a transparent outline.

Task-number: QTBUG-34855
Change-Id: I9f5c593b9093a8eb94a76081bbc2e7565c306afd
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>