Fix ASSERT on QtWebEngine TP blog post
When loading the blog post about QtWebEngine TP1, QtWebKit hits an
assert when an already loaded script stops the load of resource being
processed.

This patch changes the handling of revalidated sources so that
didAddClient and thereby notifyFinished are called from a timer event
instead of synchronously during handling of the load event.

Change-Id: I33480a7216117fa19eb6cc57d1ccac1f2886b3ae
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>