Fix assert in m_stack.isEmpty() in GraphicsContext::~GraphicsContext()
Save and restore is not balanced in GraphicsContext. This causes
the log error "ERROR void GraphicsContext::restore() stack is empty"
and assert in the GraphicsContext destructor.

The cause is a mismatch of GraphicsContexts being saved and restored by
TextureMapperImageBuffer beginClip and endClip. This patch fixes endClip
to restore the currentContext beginClip saved.

Change-Id: Ief0100dd26c7eded04734167e86712a81f1a7674
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>