Fix Qt WebKit build on architectures without JIT support
The fallback LLINT interpreter was triggering YARR JIT which would
fail to compile and also trigger ENABLE_ASSEMBLER which would trigger
even more not to compile. YARR JIT is only supported on platforms with
valid JIT or native LLINT support. So we must avoid it when the fallback
is used

Task-number: QTBUG-36969
Change-Id: Ie94bcb316e192feb33ab528ddc3a328af154cebf
Reviewed-by: Michael Bruning <michael.bruning@digia.com>
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>