Fix rendering of WebKit2 QML WebView with ANGLE.
The WebKit2 QML WebView was not rendered on ANGLE builds
due to a precision mismatch for a uniform matrix used in
in both vertex and fragment shaders which made the ANGLE
compiler bail out on linking the shader program.

This was fixed by overriding the precision to highp in
both shaders.

Task-number: QTBUG-37043
Change-Id: I840f8e348d5de1c5900576e7d68c3f8ddcdfa4a7
Reviewed-by: Simon Hausmann <simon.hausmann@digia.com>