Fix crash on access of dead QMimeData
When we give a QMimeData to the system pasteboard, we should no longer
access it locally.

This patch ensures we invalidate the pointer to the QMimeDate once it
has been uploaded to the system clipboard. To ensure we can still read
values we have writen, the code to read the pasteboard data will now
read it from the system pasteboard when applicable. Several places
the latter was done manually have been removed.

Task-number: QTBUG-37904
Change-Id: I9c37b2fcc58ebbc387319be39c190fd7db6f70eb
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>