Integrate QWebChannel with QtWebKit.
WebView.experimental gets a new webChannel property. By default, this
is initialized to a QQmlWebChannel internal to the WebView. By setting
it from the outside, it is possible to share the same WebChannel
between different WebViews. The webChannel property can be set to null
to disable this new feature.

For IPC, a navigator.qtWebChannelTransport property is added, which is
essentially a copy of navigator.qt. This is required to be able to use
both independently. The transport is implicitly connected to the
WebView's webChannel.

On the JavaScript client side, some manual boiler plate code is still
required. Potentially, this can be adapted in the future to preload
the qwebchannel.js. Furthermore, we might be able to instantiate and
initialize the QWebChannel before we emit window.onload.

A basic test is added which verifies that this basic integration works.

Change-Id: I5eb0389e6edd6d0c978917e441365e11a02b5c3f
Reviewed-by: Milian Wolff <milian.wolff@kdab.com>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@digia.com>