QWebPage crashes on drag without QWidget view
QWebPage can be used without a widgets and has a documentation section
on how to do so, but if QWegPage hasn't been given a QWidget in setView
a drag operation will crash.  To reproduce select a section of the web page
click and drag from within the selection and it will crash.

Add a check to avoid a NULL dereference, the existing code already has
a check for not having a QWidget set as the view.
DragClientQt::startDrag
m_chromeClient->platformPageClient() is NULL

Task-number: QTBUG-40204
Change-Id: I5739285696ab57c05ee59c10ae00a1cf652ff494
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@digia.com>