Added support for appending an escaped SegmentedString to another.
Merge of blink r174725

This was previously an assert which triggers in the situation where the
HTMLEntityParser is parsing a potential named entity, but only has one or two
characters of data available due to being at the end of a network packet
boundary, so it 'unconsumes' them from the SegmentedString which has some
special handling when only one or two characters needs unconsuming to avoid
string creation. Next time the tokenizer is pumped, the SegmentedString is
then appended to the HTMLSourceTracker right before the actual parsing happens
(i.e. when we'd consume the previously unconsumed characters).
SegmentedString::append mostly assumes the parameter string hasn't been
consumed from and had an assert instead of handling this case, which hits.

Task-number:  QTBUG-40284
Change-Id: Ifcf9af6f6e2e5a9750af4297cfda2f55f3bb274e
Reviewed-by: Michael Bruning <michael.bruning@digia.com>