Improve layout and rendering speed of complex font path
This patch enables the use of the TextLayout object previously only used
by the Mac port. This caches a line of laid out text making reflowing
text much faster when using the complex font path.

The patch also enables caching in the QTextLayout, this improves also
drawing and measuring of complex fonts, since previously we would throw
away details of the layout after calculating a line, but then
recalculate it for drawing or measuring.

Change-Id: I9db40cdb5a35d28072204f950a8aa50669ac643b
Reviewed-by: Pierre Rossi <pierre.rossi@gmail.com>