Need to check if some HTML child elements are HTMLUnknownElement
Based on upstream fix http://trac.webkit.org/changeset/156953

The check for whether an element is an HTMLAudioElement or not was
incomplete. An element can have the 'audio' tag-name but still be
another element if media elements have been disabled. In this
case it will be an HTMLUnknownElement.

Task-number: QTBUG-41360
Change-Id: I6a3e366b9dc268b0dbebe5880ba68945bcb42a27
Reviewed-by: Michael Bruning <michael.bruning@digia.com>
Reviewed-by: Jocelyn Turcotte <jocelyn.turcotte@digia.com>