LLInt is not built with ARM softfp ABI
In fixing QTBUG-35681 we disabled LLInt when hard fp is not set. This
is however too strong a condition as we only need to be allowed to use
floating point instructions, not use the hard floating point ABI.

LLInt is also disabled on Android because it currently doesn't build.

Task-number: QTBUG-35681
Change-Id: Ibe109b0f8fe6acaec90d42305c6151a440141ea9
Reviewed-by: Julien Brianceau <jbriance@cisco.com>
Reviewed-by: Michael Bruning <michael.bruning@digia.com>