Fix memory leak in ImageDocument
An ImageLoader may hold a reference to its image element, and any
element holds a guard on the document. To make the ImageLoader release
the reference it needs to receive the notifyFinished() call, but that
doesn't happen if the CachedResource is in the wrong state.

The solution is to ensure we set the finish state on the resource before
calling the method signifying loading is finished.

Task-number: QTBUG-44023
Change-Id: I694483f6201b7ea4656016731e431f2806c3ab1c
Reviewed-by: Michael Brüning <michael.bruning@theqtcompany.com>