Initialize label vector lazily
When parsing JSON temporary JIT objects are created when resolving each
id. Each of these get a list of labels initialized to the size of the
codeblock being operated on, which can be very long in some cases.

This patch delays the initialization of the label vector, until it is
actually used which is easy to figure out since the vector is not
exported outside the class.

Task-number: QTBUG-44475
Change-Id: I4fdbb7de7e7d953fffed39e38feed066edb6742b
Reviewed-by: Michael Brüning <michael.bruning@theqtcompany.com>