Fix build on x32 systems
On Linux x32 ABI, pointer size is different than on classic x86_64
systems, so the current assembly is wrong.

To fix this, drop the quad qualifier and let the compiler infer the
argument size. This way it will work on both x86_64 and x32.

Done-with: Adam Borowski <kilobyte@angband.pl>
Change-Id: I7a5f9edbba08705397ce7e771b1567f2a140fb1b
Reviewed-by: Thiago Macieira <thiago.macieira@intel.com>
Reviewed-by: Lisandro Damián Nicanor Pérez Meyer <perezmeyer@gmail.com>