Fix crash when exiting browser with WebGL or accelerated canvas.
Issue is that the TextureMapperGL creates a GraphicsContext3D using
the current OpenGL context. This is done by storing a reference on
QOpenGLContext::currentContext() inside
GraphicsContext3DPrivate::GraphicsContext3DPrivate(). When exiting the
browser, Qt releases QOpenGLContext before the release of the
GraphicsContext3D in webkit which leads to a crash when destoying
GraphicsContext3DPrivate.

Task-number: QTBUG-45481
Change-Id: I2b9d7b1a96fbbe8517ea323d45ef3922ada208a3
Reviewed-by: Julien Brianceau <jbriance@cisco.com>
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@theqtcompany.com>