Update the detection of MSVC versions to be forward-compatible
Qt 5.8 may deprecate the win32-msvc20xx mkspecs, so use the MSVC_VER
variable to detect which version of MSVC it is.

This commit removes the detection for win32-icc, as it really depended
on what version of MSVC headers it was targeting. With Qt prior to 5.8,
it was not possible to detect it anyway, so the rule was broken anyway
(ICC prior to 16 only worked with MSVC <= 2013, but ICC 17 will work
with 2015).

Change-Id: Ib57b52598e2f452985e9fffd1459e682978341f5
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>