Remove unused check for private_tests
It turns out that qt_developer_build is never used anyway.

I did some git archeology and it looks like it was nuked when doing the
5.2 merge from upstream. The original purpose was apparently to control
-Werror, which is nowadays controlled by CONFIG -= production_build.

This actually fixes the build, as the QT_FOR_CONFIG += core-private which
would have been required for qtConfig(private_tests) is missing.

Task-number: QTBUG-55950
Change-Id: Iaaaad184b29b523ce4a4ed8afec2ac527d8f93e3
Reviewed-by: Oswald Buddenhagen <oswald.buddenhagen@qt.io>
Reviewed-by: Konstantin Tokarev <annulen@yandex.ru>
Reviewed-by: Jan Kundrát <jkt@kde.org>