Fix build with MSYS Python
MSYS's Python is compiled as a POSIX environment, which means
subprocess.Popen cannot be passed just a single string (to be given to
CreateProcess). Instead, pass the split args and let the underlying
system re-merge them. This is harmless with a pure Windows Python.

MSYS python also changes parameters starting with / to a pathname, so
use the equivalent -directives (dumpbin.exe accepts them).

Change-Id: Id50531a20e40adb71a40525cf714d05c22959fb0
Reviewed-by: Allan Sandfeld Jensen <allan.jensen@qt.io>